Add initial GIF frames to gif_frames_history directory
This commit introduces a series of new PNG files representing the frames of a GIF animation. The frames are sequentially numbered from frame_000.png to frame_025.png, providing a complete set of images for the GIF. Each file is added with the appropriate permissions and is ready for use in the application.